requirements_dev.txt
alabaster==0.7.12
appnope==0.1.0
astroid==2.1.0
autopep8==1.4.4
Babel==2.6.0
backcall==0.1.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.7.0
cycler==0.10.0
Cython==0.29.6
decorator==4.3.2
defusedxml==0.5.0
docutils==0.14
entrypoints==0.3
greenlet==0.4.15
h5py==2.9.0
idna==2.8
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-kernel-gateway==2.1.0
keyring==17.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.28.0
MarkupSafe==1.1.0
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
msgpack==0.6.2
nbconvert==5.4.0
nbformat==4.4.0
neovim==0.3.1
notebook==5.7.4
numba==0.43.1
numpy==1.16.3
numpydoc==0.8.0
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
popen==0.1.20
prometheus-client==0.5.0
prompt-toolkit==2.0.8
psutil==5.5.0
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.0
Pygments==2.3.1
pylint==2.2.2
pynvim==0.4.0
pyparsing==2.3.1
PyQt5==5.9.2
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
pyzmq==17.1.2
QtAwesome==0.5.6
qtconsole==4.4.3
QtPy==1.6.0
requests==2.21.0
rope==0.11.0
scipy==1.2.1
Send2Trash==1.5.0
simcore-analysis==0.1.0
sip==4.19.8
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.4.2
spyder-notebook==0.1.4
style==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
update==0.0.1
urllib3==1.24.1
vtk==8.1.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
